refactor: move bedrock proxy handling to windmill-ai
