+++ backend/windmill-ai/src/providers/bedrock.rs
@@ -7,21 +7,731 @@
 //! - Helper utilities
 
 use crate::{
+    ai_bedrock::{
+        bedrock_model_supports_prompt_caching, bedrock_stream_event_is_block_stop,
+        bedrock_stream_event_to_text, bedrock_stream_event_to_tool_delta,
+        bedrock_stream_event_to_tool_start, build_tool_config, create_inference_config,
+        format_bedrock_error, openai_messages_to_bedrock, streaming_tool_calls_to_openai,
+        BearerTokenProvider, BedrockClient, StreamingToolCall,
+    },
+    ai_providers::USE_ENV_REGION,
+    ai_types::{OpenAIFunction, OpenAIToolCall, ToolDefFunction},
     image_handler::prepare_messages_for_api,
+    proxy::ProxyBuildArgs,
     query_builder::{ParsedResponse, StreamEventSink},
     types::{OpenAIMessage, StreamingEvent, TokenUsage, ToolDef},
 };
+use bytes::Bytes;
+use futures::{stream::BoxStream, StreamExt};
+use http::{HeaderMap, Method, StatusCode};
+use serde::Deserialize;
 use std::collections::HashMap;
 use windmill_common::{client::AuthedClient, error::Error};
 
-// Import shared Bedrock helpers for provider orchestration.
-use crate::ai_bedrock::{
-    bedrock_model_supports_prompt_caching, bedrock_stream_event_is_block_stop,
-    bedrock_stream_event_to_text, bedrock_stream_event_to_tool_delta,
-    bedrock_stream_event_to_tool_start, build_tool_config, create_inference_config,
-    format_bedrock_error, openai_messages_to_bedrock, streaming_tool_calls_to_openai,
-    BedrockClient, StreamingToolCall,
-};
+// ============================================================================
+// Native Proxy Execution
+// ============================================================================
+
+/// OpenAI-format request body for Bedrock SDK proxy handlers.
+#[derive(Deserialize, Debug)]
+struct OpenAIRequest {
+    messages: Vec<OpenAIMessage>,
+    #[serde(default)]
+    tools: Option<Vec<OpenAIToolDef>>,
+    #[serde(default)]
+    tool_choice: Option<serde_json::Value>,
+    #[serde(default)]
+    max_tokens: Option<i32>,
+    #[serde(default)]
+    temperature: Option<f32>,
+}
+
+#[derive(Deserialize, Debug)]
+struct OpenAIToolDef {
+    #[serde(default)]
+    #[allow(dead_code)]
+    r#type: Option<String>,
+    function: OpenAIToolFunction,
+}
+
+#[derive(Deserialize, Debug)]
+struct OpenAIToolFunction {
+    name: String,
+    #[serde(default)]
+    description: Option<String>,
+    #[serde(default)]
+    parameters: Option<serde_json::Value>,
+}
+
+#[derive(Deserialize, Debug)]
+struct BedrockProxyChatRequest {
+    model: String,
+    #[serde(default)]
+    stream: bool,
+}
+
+enum BedrockAuthConfig {
+    BearerToken(String),
+    IamCredentials {
+        access_key_id: String,
+        secret_access_key: String,
+        session_token: Option<String>,
+    },
+    Environment,
+}
+
+pub enum BedrockProxyResponseBody {
+    Fixed(Bytes),
+    Stream(BoxStream<'static, std::result::Result<Bytes, std::io::Error>>),
+}
+
+pub struct BedrockProxyResponse {
+    pub status_code: StatusCode,
+    pub headers: HeaderMap,
+    pub body: BedrockProxyResponseBody,
+}
+
+/// Handle a workspace Bedrock proxy request through the AWS SDK.
+///
+/// The API still owns credential resolution, route authorization, auditing, and
+/// cache behavior. This helper owns Bedrock-specific control-plane and
+/// OpenAI-compatible Converse transformations.
+pub async fn handle_bedrock_proxy(
+    args: &ProxyBuildArgs<'_>,
+) -> Result<BedrockProxyResponse, Error> {
+    let region = args.credentials.region.as_deref().unwrap_or(USE_ENV_REGION);
+
+    if *args.method == Method::GET {
+        return match args.path {
+            "foundation-models" => list_foundation_models(args, region).await,
+            "inference-profiles" => list_inference_profiles(args, region).await,
+            _ => Err(Error::BadRequest(format!(
+                "Unsupported AWS Bedrock proxy path: {}",
+                args.path
+            ))),
+        };
+    }
+
+    if *args.method != Method::POST {
+        return Err(Error::BadRequest(format!(
+            "Unsupported AWS Bedrock proxy method: {}",
+            args.method
+        )));
+    }
+
+    let request: BedrockProxyChatRequest = serde_json::from_slice(args.body)
+        .map_err(|e| Error::internal_err(format!("Failed to parse request body: {}", e)))?;
+
+    if request.stream {
+        handle_bedrock_sdk_streaming(&request.model, args.body, args, region).await
+    } else {
+        handle_bedrock_sdk_non_streaming(&request.model, args.body, args, region).await
+    }
+}
+
+fn determine_auth_config(
+    api_key: Option<&str>,
+    aws_access_key_id: Option<&str>,
+    aws_secret_access_key: Option<&str>,
+    aws_session_token: Option<&str>,
+) -> BedrockAuthConfig {
+    if let Some(key) = api_key.filter(|k| !k.is_empty()) {
+        BedrockAuthConfig::BearerToken(key.to_string())
+    } else if let (Some(access_key_id), Some(secret_access_key)) = (
+        aws_access_key_id.filter(|s| !s.is_empty()),
+        aws_secret_access_key.filter(|s| !s.is_empty()),
+    ) {
+        BedrockAuthConfig::IamCredentials {
+            access_key_id: access_key_id.to_string(),
+            secret_access_key: secret_access_key.to_string(),
+            session_token: aws_session_token
+                .filter(|token| !token.is_empty())
+                .map(str::to_string),
+        }
+    } else {
+        BedrockAuthConfig::Environment
+    }
+}
+
+async fn create_bedrock_client(
+    args: &ProxyBuildArgs<'_>,
+    region: &str,
+) -> Result<BedrockClient, Error> {
+    match determine_auth_config(
+        args.credentials.api_key.as_deref(),
+        args.credentials.aws_access_key_id.as_deref(),
+        args.credentials.aws_secret_access_key.as_deref(),
+        args.credentials.aws_session_token.as_deref(),
+    ) {
+        BedrockAuthConfig::BearerToken(key) => BedrockClient::from_bearer_token(key, region).await,
+        BedrockAuthConfig::IamCredentials { access_key_id, secret_access_key, session_token } => {
+            BedrockClient::from_credentials(access_key_id, secret_access_key, session_token, region)
+                .await
+        }
+        BedrockAuthConfig::Environment => BedrockClient::from_env(region).await,
+    }
+}
+
+fn build_tool_config_from_request(
+    tools: Option<&[OpenAIToolDef]>,
+    tool_choice: Option<&serde_json::Value>,
+    enable_prompt_caching: bool,
+) -> Result<Option<aws_sdk_bedrockruntime::types::ToolConfiguration>, Error> {
+    if let Some(tools) = tools {
+        let tool_defs: Vec<ToolDef> = tools
+            .iter()
+            .map(|t| ToolDef {
+                r#type: "function".to_string(),
+                function: ToolDefFunction {
+                    name: t.function.name.clone(),
+                    description: t.function.description.clone(),
+                    parameters: Box::from(
+                        serde_json::value::RawValue::from_string(
+                            serde_json::to_string(
+                                &t.function
+                                    .parameters
+                                    .clone()
+                                    .unwrap_or(serde_json::json!({})),
+                            )
+                            .unwrap_or_default(),
+                        )
+                        .unwrap_or_else(|_| {
+                            serde_json::value::RawValue::from_string("{}".to_string()).unwrap()
+                        }),
+                    ),
+                },
+            })
+            .collect();
+
+        let force_tool_use = tool_choice
+            .map(|tc| tc == "required" || tc.as_str() == Some("required"))
+            .unwrap_or(false);
+
+        build_tool_config(Some(&tool_defs), force_tool_use, enable_prompt_caching)
+    } else {
+        Ok(None)
+    }
+}
+
+async fn create_bedrock_control_client(
+    args: &ProxyBuildArgs<'_>,
+    region: &str,
+) -> Result<aws_sdk_bedrock::Client, Error> {
+    use aws_config::BehaviorVersion;
+
+    let region_provider = aws_sdk_bedrock::config::Region::new(region.to_string());
+
+    match determine_auth_config(
+        args.credentials.api_key.as_deref(),
+        args.credentials.aws_access_key_id.as_deref(),
+        args.credentials.aws_secret_access_key.as_deref(),
+        args.credentials.aws_session_token.as_deref(),
+    ) {
+        BedrockAuthConfig::BearerToken(key) => {
+            let config = aws_sdk_bedrock::config::Builder::new()
+                .region(region_provider)
+                .behavior_version(BehaviorVersion::latest())
+                .token_provider(BearerTokenProvider::new(key))
+                .build();
+            Ok(aws_sdk_bedrock::Client::from_conf(config))
+        }
+        BedrockAuthConfig::IamCredentials { access_key_id, secret_access_key, session_token } => {
+            let credentials = aws_credential_types::Credentials::new(
+                access_key_id,
+                secret_access_key,
+                session_token,
+                None,
+                "windmill",
+            );
+            let config = aws_sdk_bedrock::config::Builder::new()
+                .region(region_provider)
+                .behavior_version(BehaviorVersion::latest())
+                .credentials_provider(credentials)
+                .build();
+            Ok(aws_sdk_bedrock::Client::from_conf(config))
+        }
+        BedrockAuthConfig::Environment => {
+            let config = aws_config::defaults(BehaviorVersion::latest())
+                .region(region_provider)
+                .load()
+                .await;
+            Ok(aws_sdk_bedrock::Client::new(&config))
+        }
+    }
+}
+
+async fn list_foundation_models(
+    args: &ProxyBuildArgs<'_>,
+    region: &str,
+) -> Result<BedrockProxyResponse, Error> {
+    let client = create_bedrock_control_client(args, region).await?;
+
+    let response = client
+        .list_foundation_models()
+        .send()
+        .await
+        .map_err(|e| Error::internal_err(format!("Failed to list foundation models: {}", e)))?;
+
+    let models: Vec<serde_json::Value> = response
+        .model_summaries()
+        .iter()
+        .map(|m| {
+            serde_json::json!({
+                "modelId": m.model_id(),
+                "modelName": m.model_name(),
+                "providerName": m.provider_name(),
+                "modelArn": m.model_arn(),
+                "inputModalities": m.input_modalities().iter().map(|i| i.as_str()).collect::<Vec<_>>(),
+                "outputModalities": m.output_modalities().iter().map(|o| o.as_str()).collect::<Vec<_>>(),
+                "responseStreamingSupported": m.response_streaming_supported(),
+                "inferenceTypesSupported": m.inference_types_supported().iter().map(|i| i.as_str()).collect::<Vec<_>>(),
+            })
+        })
+        .collect();
+
+    let body = serde_json::to_vec(&serde_json::json!({ "modelSummaries": models }))
+        .map_err(|e| Error::internal_err(format!("Failed to serialize response: {}", e)))?;
+
+    Ok(BedrockProxyResponse {
+        status_code: StatusCode::OK,
+        headers: json_response_headers(),
+        body: BedrockProxyResponseBody::Fixed(Bytes::from(body)),
+    })
+}
+
+async fn list_inference_profiles(
+    args: &ProxyBuildArgs<'_>,
+    region: &str,
+) -> Result<BedrockProxyResponse, Error> {
+    let client = create_bedrock_control_client(args, region).await?;
+
+    let response =
+        client.list_inference_profiles().send().await.map_err(|e| {
+            Error::internal_err(format!("Failed to list inference profiles: {}", e))
+        })?;
+
+    let profiles: Vec<serde_json::Value> = response
+        .inference_profile_summaries()
+        .iter()
+        .map(|p| {
+            serde_json::json!({
+                "inferenceProfileId": p.inference_profile_id(),
+                "inferenceProfileName": p.inference_profile_name(),
+                "inferenceProfileArn": p.inference_profile_arn(),
+                "description": p.description(),
+                "status": p.status().as_str(),
+                "type": p.r#type().as_str(),
+            })
+        })
+        .collect();
+
+    let body = serde_json::to_vec(&serde_json::json!({ "inferenceProfileSummaries": profiles }))
+        .map_err(|e| Error::internal_err(format!("Failed to serialize response: {}", e)))?;
+
+    Ok(BedrockProxyResponse {
+        status_code: StatusCode::OK,
+        headers: json_response_headers(),
+        body: BedrockProxyResponseBody::Fixed(Bytes::from(body)),
+    })
+}
+
+async fn handle_bedrock_sdk_streaming(
+    model: &str,
+    body: &[u8],
+    args: &ProxyBuildArgs<'_>,
+    region: &str,
+) -> Result<BedrockProxyResponse, Error> {
+    let openai_req: OpenAIRequest = serde_json::from_slice(body)
+        .map_err(|e| Error::internal_err(format!("Failed to parse OpenAI request: {}", e)))?;
+
+    let bedrock_client = create_bedrock_client(args, region).await?;
+    let enable_prompt_caching = bedrock_model_supports_prompt_caching(model);
+    let (bedrock_messages, system_prompts) =
+        openai_messages_to_bedrock(&openai_req.messages, enable_prompt_caching)?;
+    let inference_config = create_inference_config(openai_req.temperature, openai_req.max_tokens);
+    let tool_config = build_tool_config_from_request(
+        openai_req.tools.as_deref(),
+        openai_req.tool_choice.as_ref(),
+        enable_prompt_caching,
+    )?;
+
+    let mut request_builder = bedrock_client
+        .client()
+        .converse_stream()
+        .model_id(model)
+        .set_messages(Some(bedrock_messages));
+
+    if !system_prompts.is_empty() {
+        request_builder = request_builder.set_system(Some(system_prompts));
+    }
+
+    if let Some(config) = inference_config {
+        request_builder = request_builder.inference_config(config);
+    }
+
+    if let Some(config) = tool_config {
+        request_builder = request_builder.set_tool_config(Some(config));
+    }
+
+    tracing::debug!("Bedrock SDK streaming: sending converse_stream request");
+    let stream_output = request_builder.send().await.map_err(|e| {
+        let error_msg = format!("Bedrock SDK streaming error: {}", format_bedrock_error(&e));
+        tracing::error!("Bedrock SDK streaming failed: {}", error_msg);
+        Error::internal_err(error_msg)
+    })?;
+    tracing::debug!("Bedrock SDK streaming: stream established successfully");
+
+    Ok(BedrockProxyResponse {
+        status_code: StatusCode::OK,
+        headers: event_stream_response_headers(),
+        body: BedrockProxyResponseBody::Stream(
+            sdk_stream_to_sse(stream_output.stream, model.to_string()).boxed(),
+        ),
+    })
+}
+
+pub fn sdk_stream_to_sse(
+    stream: aws_sdk_bedrockruntime::primitives::event_stream::EventReceiver<
+        aws_sdk_bedrockruntime::types::ConverseStreamOutput,
+        aws_sdk_bedrockruntime::types::error::ConverseStreamOutputError,
+    >,
+    model: String,
+) -> impl futures::Stream<Item = std::result::Result<Bytes, std::io::Error>> + Send {
+    let id = format!("chatcmpl-{}", uuid::Uuid::new_v4().simple());
+    let created = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .unwrap()
+        .as_secs();
+
+    struct StreamState {
+        id: String,
+        model: String,
+        created: u64,
+        tool_calls: HashMap<usize, (String, String, String)>,
+        current_tool_index: usize,
+    }
+
+    let state = std::sync::Arc::new(tokio::sync::Mutex::new(StreamState {
+        id,
+        model,
+        created,
+        tool_calls: HashMap::new(),
+        current_tool_index: 0,
+    }));
+
+    async_stream::stream! {
+        let mut stream = stream;
+        let state = state.clone();
+
+        loop {
+            match stream.recv().await {
+                Ok(Some(event)) => {
+                    let mut state = state.lock().await;
+
+                    if let Some(tool_call) = bedrock_stream_event_to_tool_start(&event) {
+                        let index = state.current_tool_index;
+                        state.tool_calls.insert(
+                            index,
+                            (tool_call.id.clone(), tool_call.name.clone(), String::new()),
+                        );
+
+                        let chunk = serde_json::json!({
+                            "id": state.id,
+                            "object": "chat.completion.chunk",
+                            "created": state.created,
+                            "model": state.model,
+                            "choices": [{
+                                "index": 0,
+                                "delta": {
+                                    "tool_calls": [{
+                                        "index": index,
+                                        "id": tool_call.id,
+                                        "type": "function",
+                                        "function": {
+                                            "name": tool_call.name,
+                                            "arguments": ""
+                                        }
+                                    }]
+                                },
+                                "finish_reason": serde_json::Value::Null
+                            }]
+                        });
+
+                        yield Ok(Bytes::from(format!("data: {}\n\n", chunk)));
+                    }
+
+                    if let Some(text) = bedrock_stream_event_to_text(&event) {
+                        let chunk = serde_json::json!({
+                            "id": state.id,
+                            "object": "chat.completion.chunk",
+                            "created": state.created,
+                            "model": state.model,
+                            "choices": [{
+                                "index": 0,
+                                "delta": {
+                                    "content": text
+                                },
+                                "finish_reason": serde_json::Value::Null
+                            }]
+                        });
+
+                        yield Ok(Bytes::from(format!("data: {}\n\n", chunk)));
+                    }
+
+                    if let Some(input_delta) = bedrock_stream_event_to_tool_delta(&event) {
+                        let index = state.current_tool_index;
+                        if let Some((_id, _name, ref mut args)) = state.tool_calls.get_mut(&index) {
+                            args.push_str(&input_delta);
+
+                            let chunk = serde_json::json!({
+                                "id": state.id,
+                                "object": "chat.completion.chunk",
+                                "created": state.created,
+                                "model": state.model,
+                                "choices": [{
+                                    "index": 0,
+                                    "delta": {
+                                        "tool_calls": [{
+                                            "index": index,
+                                            "function": {
+                                                "arguments": input_delta
+                                            }
+                                        }]
+                                    },
+                                    "finish_reason": serde_json::Value::Null
+                                }]
+                            });
+
+                            yield Ok(Bytes::from(format!("data: {}\n\n", chunk)));
+                        }
+                    }
+
+                    if bedrock_stream_event_is_block_stop(&event) {
+                        state.current_tool_index += 1;
+                    }
+
+                    if let aws_sdk_bedrockruntime::types::ConverseStreamOutput::MessageStop(stop) = &event {
+                        let stop_reason = stop.stop_reason().as_str();
+                        let finish_reason = match stop_reason {
+                            "end_turn" => "stop",
+                            "max_tokens" => "length",
+                            "tool_use" => "tool_calls",
+                            "stop_sequence" => "stop",
+                            "guardrail_intervened" | "content_filtered" => "content_filter",
+                            _ => "stop",
+                        };
+
+                        let chunk = serde_json::json!({
+                            "id": state.id,
+                            "object": "chat.completion.chunk",
+                            "created": state.created,
+                            "model": state.model,
+                            "choices": [{
+                                "index": 0,
+                                "delta": {},
+                                "finish_reason": finish_reason
+                            }]
+                        });
+
+                        yield Ok(Bytes::from(format!("data: {}\n\n", chunk)));
+                    }
+                }
+                Ok(None) => break,
+                Err(e) => {
+                    yield Err(std::io::Error::new(
+                        std::io::ErrorKind::Other,
+                        e.to_string(),
+                    ));
+                    break;
+                }
+            }
+        }
+
+        yield Ok(Bytes::from("data: [DONE]\n\n"));
+    }
+}
+
+async fn handle_bedrock_sdk_non_streaming(
+    model: &str,
+    body: &[u8],
+    args: &ProxyBuildArgs<'_>,
+    region: &str,
+) -> Result<BedrockProxyResponse, Error> {
+    let openai_req: OpenAIRequest = serde_json::from_slice(body)
+        .map_err(|e| Error::internal_err(format!("Failed to parse OpenAI request: {}", e)))?;
+
+    let bedrock_client = create_bedrock_client(args, region).await?;
+    let enable_prompt_caching = bedrock_model_supports_prompt_caching(model);
+    let (bedrock_messages, system_prompts) =
+        openai_messages_to_bedrock(&openai_req.messages, enable_prompt_caching)?;
+    let inference_config = create_inference_config(openai_req.temperature, openai_req.max_tokens);
+    let tool_config = build_tool_config_from_request(
+        openai_req.tools.as_deref(),
+        openai_req.tool_choice.as_ref(),
+        enable_prompt_caching,
+    )?;
+
+    let mut request_builder = bedrock_client
+        .client()
+        .converse()
+        .model_id(model)
+        .set_messages(Some(bedrock_messages));
+
+    if !system_prompts.is_empty() {
+        request_builder = request_builder.set_system(Some(system_prompts));
+    }
+
+    if let Some(config) = inference_config {
+        request_builder = request_builder.inference_config(config);
+    }
+
+    if let Some(config) = tool_config {
+        request_builder = request_builder.set_tool_config(Some(config));
+    }
+
+    tracing::debug!("Bedrock SDK non-streaming: sending converse request");
+    let response = request_builder.send().await.map_err(|e| {
+        let error_msg = format!(
+            "Bedrock SDK non-streaming error: {}",
+            format_bedrock_error(&e)
+        );
+        tracing::error!("Bedrock SDK non-streaming failed: {}", error_msg);
+        Error::internal_err(error_msg)
+    })?;
+    tracing::debug!(
+        "Bedrock SDK non-streaming: response received, stop_reason={}",
+        response.stop_reason().as_str()
+    );
+
+    let id = format!("chatcmpl-{}", uuid::Uuid::new_v4().simple());
+    let created = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .unwrap()
+        .as_secs();
+
+    let stop_reason = response.stop_reason().as_str();
+    let finish_reason = match stop_reason {
+        "end_turn" => "stop",
+        "max_tokens" => "length",
+        "tool_use" => "tool_calls",
+        "stop_sequence" => "stop",
+        "guardrail_intervened" | "content_filtered" => "content_filter",
+        _ => "stop",
+    };
+
+    let mut text_content = String::new();
+    let mut tool_calls: Vec<OpenAIToolCall> = Vec::new();
+
+    if let Some(aws_sdk_bedrockruntime::types::ConverseOutput::Message(message)) = response.output()
+    {
+        for block in message.content() {
+            match block {
+                aws_sdk_bedrockruntime::types::ContentBlock::Text(text) => {
+                    text_content.push_str(text);
+                }
+                aws_sdk_bedrockruntime::types::ContentBlock::ToolUse(tool_use) => {
+                    let input_json = document_to_json(tool_use.input());
+                    tool_calls.push(OpenAIToolCall {
+                        id: tool_use.tool_use_id().to_string(),
+                        function: OpenAIFunction {
+                            name: tool_use.name().to_string(),
+                            arguments: serde_json::to_string(&input_json).unwrap_or_default(),
+                        },
+                        r#type: "function".to_string(),
+                        extra_content: None,
+                    });
+                }
+                _ => {}
+            }
+        }
+    }
+
+    let message = if !tool_calls.is_empty() {
+        serde_json::json!({
+            "role": "assistant",
+            "content": if text_content.is_empty() { serde_json::Value::Null } else { serde_json::Value::String(text_content) },
+            "tool_calls": tool_calls
+        })
+    } else {
+        serde_json::json!({
+            "role": "assistant",
+            "content": text_content
+        })
+    };
+
+    let usage = if let Some(usage_data) = response.usage() {
+        serde_json::json!({
+            "prompt_tokens": usage_data.input_tokens(),
+            "completion_tokens": usage_data.output_tokens(),
+            "total_tokens": usage_data.total_tokens()
+        })
+    } else {
+        serde_json::json!({
+            "prompt_tokens": 0,
+            "completion_tokens": 0,
+            "total_tokens": 0
+        })
+    };
+
+    let openai_resp = serde_json::json!({
+        "id": id,
+        "object": "chat.completion",
+        "created": created,
+        "model": model,
+        "choices": [{
+            "index": 0,
+            "message": message,
+            "finish_reason": finish_reason
+        }],
+        "usage": usage
+    });
+
+    let body = serde_json::to_vec(&openai_resp)
+        .map_err(|e| Error::internal_err(format!("Failed to serialize OpenAI response: {}", e)))?;
+
+    Ok(BedrockProxyResponse {
+        status_code: StatusCode::OK,
+        headers: json_response_headers(),
+        body: BedrockProxyResponseBody::Fixed(Bytes::from(body)),
+    })
+}
+
+fn json_response_headers() -> HeaderMap {
+    let mut headers = HeaderMap::new();
+    headers.insert("content-type", "application/json".parse().unwrap());
+    headers
+}
+
+fn event_stream_response_headers() -> HeaderMap {
+    let mut headers = HeaderMap::new();
+    headers.insert("content-type", "text/event-stream".parse().unwrap());
+    headers.insert("cache-control", "no-cache".parse().unwrap());
+    headers.insert("connection", "keep-alive".parse().unwrap());
+    headers
+}
+
+fn document_to_json(doc: &aws_smithy_types::Document) -> serde_json::Value {
+    match doc {
+        aws_smithy_types::Document::Object(map) => {
+            let mut json_map = serde_json::Map::new();
+            for (key, value) in map {
+                json_map.insert(key.clone(), document_to_json(value));
+            }
+            serde_json::Value::Object(json_map)
+        }
+        aws_smithy_types::Document::Array(values) => {
+            serde_json::Value::Array(values.iter().map(document_to_json).collect())
+        }
+        aws_smithy_types::Document::Number(number) => match number {
+            aws_smithy_types::Number::PosInt(number) => serde_json::Value::Number((*number).into()),
+            aws_smithy_types::Number::NegInt(number) => serde_json::Value::Number((*number).into()),
+            aws_smithy_types::Number::Float(number) => serde_json::json!(*number),
+        },
+        aws_smithy_types::Document::String(value) => serde_json::Value::String(value.clone()),
+        aws_smithy_types::Document::Bool(value) => serde_json::Value::Bool(*value),
+        aws_smithy_types::Document::Null => serde_json::Value::Null,
+    }
+}
 
 // ============================================================================
 // Query Builder
@@ -254,5 +964,62 @@ impl BedrockQueryBuilder {
             used_websearch: false,
             usage,
         })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn determine_auth_config_prioritizes_bearer_token() {
+        let config = determine_auth_config(
+            Some("bearer-token"),
+            Some("AKIA123"),
+            Some("secret"),
+            Some("session-token"),
+        );
+
+        match config {
+            BedrockAuthConfig::BearerToken(token) => assert_eq!(token, "bearer-token"),
+            _ => panic!("expected bearer token auth config"),
+        }
+    }
+
+    #[test]
+    fn determine_auth_config_uses_iam_with_optional_session_token() {
+        let config =
+            determine_auth_config(None, Some("AKIA123"), Some("secret"), Some("session-token"));
+
+        match config {
+            BedrockAuthConfig::IamCredentials {
+                access_key_id,
+                secret_access_key,
+                session_token,
+            } => {
+                assert_eq!(access_key_id, "AKIA123");
+                assert_eq!(secret_access_key, "secret");
+                assert_eq!(session_token.as_deref(), Some("session-token"));
+            }
+            _ => panic!("expected IAM auth config"),
+        }
+    }
+
+    #[test]
+    fn determine_auth_config_treats_empty_session_token_as_none() {
+        let config = determine_auth_config(None, Some("AKIA123"), Some("secret"), Some(""));
+
+        match config {
+            BedrockAuthConfig::IamCredentials { session_token, .. } => {
+                assert!(session_token.is_none());
+            }
+            _ => panic!("expected IAM auth config"),
+        }
+    }
+
+    #[test]
+    fn determine_auth_config_falls_back_to_environment() {
+        let config = determine_auth_config(None, Some("AKIA123"), None, Some("session-token"));
+        assert!(matches!(config, BedrockAuthConfig::Environment));
     }
 }
